data/requirements.txt
rasterio
numpy
matplotlib
PyYAML
tdqm
opencv-python
--index-url https://download.pytorch.org/whl/cu118
torch
torchvision